requirements.txt
aiogram==3.15.0
groq>=0.13.0
google-generativeai==0.8.3
aiosqlite==0.20.0
openpyxl==3.1.5
reportlab==4.2.2
python-dotenv==1.0.1
matplotlib==3.9.3
Pillow==11.0.0
